implement support for k-layers
Support for k-layer neural network has been added.
The computation of cost and gradients have been adjusted accordingly.

When initiating the network it is now possible to specify the number of hidden layers and the number of nodes for each layer using  the "hidden_layer="-argument.

Takes a list where each entry is the number of nodes in that layer.

Ex.
hidden_layer=[100, 50, 10] would initiate a network with 3 hidden layers with 100, 50 and 10 nodes respectively.